apps/api/pyproject.toml
[project]
name = "investiq-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.29.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "pydantic[email]>=2.7.0",
  "pydantic-settings>=2.2.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "cryptography>=42.0.0",
  "redis>=5.0.0",
  "httpx>=0.27.0",
  "yfinance>=0.2.40",
  "pandas>=2.2.0",
  "apscheduler>=3.10.4",
  "tzdata>=2024.1",
  "slowapi>=0.1.9",
  "resend>=0.8.0",
  "anthropic>=0.28.0",
  "openai>=1.30.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.27.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
